Review/bundlingminification (#3)
* Review changes

1. Removed empty document check
2. Removed page handler type check
3. Removed page backupup copy usage and try catch as well as it seems unnecessary

* Renaming a folder
