--- Composite.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.sln
@@ -1,38 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29519.181
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orckestra.Web.BundlingAndMinification", "Orckestra.Web.BundlingAndMinification\Orckestra.Web.BundlingAndMinification.csproj", "{11CC9B2D-EA0E-40A5-950D-B7FB013BAA55}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orckestra.Web.BundlingAndMinification.Package", "Orckestra.Web.BundlingAndMinification.Package\Orckestra.Web.BundlingAndMinification.Package.csproj", "{5D77A993-FE99-47E2-B896-2E5F0C9A7A5F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orckestra.Web.Css.CompileFoundation", "Orckestra.Web.Css.CompileFoundation\Orckestra.Web.Css.CompileFoundation.csproj", "{CFA12325-8684-4E00-81EB-D66B06B17B66}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{11CC9B2D-EA0E-40A5-950D-B7FB013BAA55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{11CC9B2D-EA0E-40A5-950D-B7FB013BAA55}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{11CC9B2D-EA0E-40A5-950D-B7FB013BAA55}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{11CC9B2D-EA0E-40A5-950D-B7FB013BAA55}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5D77A993-FE99-47E2-B896-2E5F0C9A7A5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5D77A993-FE99-47E2-B896-2E5F0C9A7A5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5D77A993-FE99-47E2-B896-2E5F0C9A7A5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5D77A993-FE99-47E2-B896-2E5F0C9A7A5F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CFA12325-8684-4E00-81EB-D66B06B17B66}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CFA12325-8684-4E00-81EB-D66B06B17B66}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CFA12325-8684-4E00-81EB-D66B06B17B66}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CFA12325-8684-4E00-81EB-D66B06B17B66}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9AB0C92B-4CA1-4CF1-8E48-27F96C78D84F}
-		EnterpriseLibraryConfigurationToolBinariesPathV6 = packages\EnterpriseLibrary.Common.6.0.1304.0\lib\NET45;packages\EnterpriseLibrary.Validation.6.0.1304.0\lib\NET45
-	EndGlobalSection
-EndGlobal
