+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.Package/Orckestra.Web.BundlingAndMinification.Package.csproj
