+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.Package/Package/install.xml
