+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.Package/Properties/AssemblyInfo.cs
