+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.Package/Web.Debug.config
