+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification.Package/Web.config
