+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification/Properties/AssemblyInfo.cs
