+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.BundlingAndMinification/packages.config
