+++ Orckestra.Web.BundlingAndMinification/Orckestra.Web.Css.CompileFoundation/Properties/AssemblyInfo.cs
